requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.7.0
django_celery_results==2.6.0
djangorestframework==3.16.0
gevent==25.5.1
greenlet==3.2.3
humanize==4.12.3
idna==3.10
kombu==5.5.4
msgpack==1.1.1
oauthlib==3.2.2
packaging==25.0
paramiko==3.5.1
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
python-a2s==1.4.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.4.4
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
